+++ configs/classification/cifar100/samix/auto_train_cifar_samix.py
@@ -15,10 +15,10 @@ def main():
             |   |   |--> ...
     """
 
-    # *** default CE ***
-    base_path = "configs/classification/cifar100/automix/basic/r18_l2_a2_near_lam_cat_L1_01_mlr5e_2.py"
-    # base_path = "configs/classification/cifar100/automix/basic/rx50_l2_a2_near_lam_cat_L1_01_mlr5e_2.py"
-    # base_path = "configs/classification/cifar100/automix/basic/wrn28_8_l1_a2_near_lam_cat_L1_01_mlr1e_3.py"
+    ## basic
+    base_path = "configs/classification/cifar100/samix/basic/r18_l2_a2_bili_val_dp0_mul_x_cat_L1_var_01_mlr5e_2.py"
+    # base_path = "configs/classification/cifar100/samix/basic/rx50_l2_a2_bili_val_dp0_mul_x_cat_L1_var_01_mlr5e_2.py"
+    # base_path = "configs/classification/cifar100/samix/basic/wrn28_8_l1_a2_bili_val_dp0_mul_x_cat_L1_var_01_mlr1e_3.py"
 
     # abbreviation of long attributes
     abbs = {
@@ -26,13 +26,15 @@ def main():
     }
     # create nested dirs (cannot be none)
     model_var = {
-        'model.mix_block.unsampling_mode': ['nearest',],
+        'model.mix_block.unsampling_mode': ['bilinear',],
     }
     # adjust sub-attributes (cannot be none)
     gm_var = {
         # 'model.alpha': [2,],  # default: 2
         # 'model.head_mix.loss.use_soft': [True, ],  # soft CE for bb cls
         # 'model.head_mix.loss.use_sigmoid': [True, ],  # BCE for bb cls
+        # 'model.mask_adjust': [0, 0.25, 0.50,],  # SAMix, small datasets, default: 0
+        # 'model.mix_block.lam_mul_k': [-1, 0.25, 0.5, 1],  # SAMix, small datasets, default: -1
         # # 'lr_config.min_lr': [5e-2, 1e-3, 1e-4, 0],  # AutoMix default: 5e-2
         'total_epochs': [400, 800, 1200]
     }