+++ configs/classification/tiny_imagenet/mixups/basic/r18_mixups_CE_none_multi_mode.py
@@ -1,0 +1,85 @@
+_base_ = '../../../../base.py'
+# model settings
+model = dict(
+    type='MixUpClassification',
+    pretrained=None,
+    alpha=[0.1, 1, 1,],  # list of alpha
+    mix_mode=["mixup", "cutmix", "vanilla",],  # list of chosen mixup modes
+    mix_prob=None,  # list of applying probs (sum=1), None for random applying
+    mix_repeat=1,  # times of repeating mixup aug
+    mix_args=dict(
+        manifoldmix=dict(layer=(0, 3)),
+        resizemix=dict(scope=(0.1, 0.8), use_alpha=False),
+        fmix=dict(decay_power=3, size=(64,64), max_soft=0., reformulate=False)
+    ),
+    backbone=dict(
+        type='ResNet_CIFAR',  # CIFAR version
+        # type='ResNet_Mix_CIFAR',  # required by 'manifoldmix'
+        depth=18,
+        num_stages=4,
+        out_indices=(3,),  # no conv-1, x-1: stage-x
+        style='pytorch'),
+    head=dict(
+        type='ClsHead',  # normal CE loss
+        loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
+        with_avg_pool=True, multi_label=False, in_channels=512, num_classes=200)
+)
+# dataset settings
+data_source_cfg = dict(type='ImageNet')
+# Tiny Imagenet
+data_train_list = 'data/TinyImageNet/meta/train_labeled.txt'  # train 10w
+data_train_root = 'data/TinyImageNet/train/'
+data_test_list = 'data/TinyImageNet/meta/val_labeled.txt'  # val 1w
+data_test_root = 'data/TinyImageNet/val/'
+
+dataset_type = 'ClassificationDataset'
+img_norm_cfg = dict(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
+train_pipeline = [
+    dict(type='RandomResizedCrop', size=64),
+    dict(type='RandomHorizontalFlip'),
+]
+test_pipeline = []
+# prefetch
+prefetch = True
+if not prefetch:
+    train_pipeline.extend([dict(type='ToTensor'), dict(type='Normalize', **img_norm_cfg)])
+test_pipeline.extend([dict(type='ToTensor'), dict(type='Normalize', **img_norm_cfg)])
+
+data = dict(
+    imgs_per_gpu=100,
+    workers_per_gpu=4,
+    train=dict(
+        type=dataset_type,
+        data_source=dict(
+            list_file=data_train_list, root=data_train_root,
+            **data_source_cfg),
+        pipeline=train_pipeline,
+        prefetch=prefetch,
+    ),
+    val=dict(
+        type=dataset_type,
+        data_source=dict(
+            list_file=data_test_list, root=data_test_root, **data_source_cfg),
+        pipeline=test_pipeline,
+        prefetch=False,
+    ))
+# additional hooks
+custom_hooks = [
+    dict(type='ValidateHook',
+        dataset=data['val'],
+        initial=False,
+        interval=1,
+        imgs_per_gpu=100,
+        workers_per_gpu=4,
+        eval_param=dict(topk=(1, 5)))
+]
+# optimizer
+optimizer = dict(type='SGD', lr=0.2, momentum=0.9, weight_decay=0.0001)
+optimizer_config = dict(grad_clip=None)
+
+# lr scheduler
+lr_config = dict(policy='CosineAnnealing', min_lr=0)
+checkpoint_config = dict(interval=800)
+
+# runtime settings
+total_epochs = 400