+++ openmixup/utils/__init__.py
@@ -1,7 +1,7 @@
 from .alias_multinomial import AliasMethod
 from .collect import nondist_forward_collect, dist_forward_collect
 from .collect_env import collect_env
-from .config_tools import traverse_replace
+from .config_tools import ConfigGenerator, traverse_replace
 from .dist_utils import (allreduce_grads, allreduce_params, get_dist_info,
                          init_dist, master_only)
 from .fp16_utils import LossScaler, auto_fp16, force_fp32, wrap_fp16_model